hudson-behavior: switching to Behaviour.specify
Note: there are places where the original code was relying on whitespace in selectors
as significant to avoid collisions -- when two selectors collided, Behaviour.specify
would drop the older instances.

Instead of doing that, we assign each selector a unique ID.

We also assign each selector a linear priority, which was more or less happening
automatically, minus the above gotcha.